Fixed pretty app: didn't work with new parallel bvh build code.

+++ bvh.h
@@ -83,7 +83,7 @@ public:
 private:
 	void Subdivide( uint nodeIdx, uint depth, uint& nodePtr, float3& centroidMin, float3& centroidMax );
 	void UpdateNodeBounds( uint nodeIdx, float3& centroidMin, float3& centroidMax );
-	float FindBestSplitPlane( BVHNode& node, int& axis, float& splitPos, float3& centroidMin, float3& centroidMax );
+	float FindBestSplitPlane( BVHNode& node, int& axis, int& splitPos, float3& centroidMin, float3& centroidMax );
 	class Mesh* mesh = 0;
 public:
 	uint* triIdx = 0;
